Add support for a Grafana Loki output
Graylog is limited in what you plot on graphs and dashboards. But Grafana is amazing. That said, Grafana LogQl is not as simple as Graylog for simply querying the logs.
This commit aims to be able to just run both systems side-by-side and split logs to a Grafana Loki installation along side the Graylog outputs.

Logs coming from docker are keyed with things like "source.env". Graylog converts these sorts of keys to "source_env" before saving to the index. For logs going to Loki, we should format it the same.